split off ValuesInterface
